requirements.txt
backports-abc
beautifulsoup4
bs4
future
futures
markdown
markdown2
pyparsing
python-dateutil
PyYAML
singledispatch
six
tornado
gitpython
requests
prettierfier
jsonpickle
weasyprint
pypandoc
cssutils
